% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrosophilaData.R
\docType{data}
\name{MaleViscera}
\alias{MaleViscera}
\title{Gene expression (bulk RNA-seq) form male viscera of nine Drosophila species}
\format{
A data frame with 8591 rows and 38 columns. First two columns
\code{FBgnDmel}, and \code{SymbolDmel} give the flybase gene ids and symbols
of the 8591 orthologs in D. melanogaster. The rest of the 36 columns in the
form \code{M_VS_<species>_<replicate>} give the normalized gene expression
of each ortholog in 4 replicates of each of the 9 species.
}
\source{
\url{https://doi.org/10.26508/lsa.201800156}
}
\usage{
MaleViscera
}
\description{
Replicated gene expression (bulk RNA-seq) data from the male viscera of the
following 9 drosophila species (in 4 replicates: R1, R2, R3, R4):
  D. melanogaster (dmel)
  D. yakuba (dyak)
  D. ananassae (dana)
  D. pseudoobscura (dpse)
  D. persimilis (dper)
  D. willistoni (dwil)
  D. virilis (dvir)
  D. mojavensis (dmoj)
  D. grimshawi (dgri)
}
\keyword{datasets}
